Created "Logger - Albatross v0.0"
Overview
========
This logger for Python 3 is used to for simple debugging, error
messaging, and flags that are useful for software engineering. This
logger is fairly lightweight and versalite, allowing for different
"policies" to be given for logging on the console, file, or other custom
policies.

Current Tasks
=============
This logger was a small, one day build, meaning it can go in several
directions for advancement. Such options include:

1. Testing; There are currently no unit tests done for this logger and
should be pressed on before continuing.
2. More policies, including outputting to multiple policies
(MultiPolicy)
3. Reporting Level; When should the logger not report/print? The logger
shouldn't print an error file when using an actual build.
4. Graphical User Interface
5. Multi-threading(?) (This may be left with the Policies instead of the
logger)

Developer Goals
===============
The goal of this project was just to make a useful tool and have more
practice on practicality and programming. I found the experience to be
enjoyable and hope to use this for future projects.